+++ crates/boatramp-server/src/lib.rs
@@ -74,7 +74,11 @@ mod graphql_apq;
 #[cfg(feature = "handlers")]
 mod graphql_federation;
 #[cfg(feature = "handlers")]
+mod graphql_gateway;
+#[cfg(feature = "handlers")]
 mod graphql_guard;
+#[cfg(feature = "handlers")]
+mod graphql_plan;
 #[cfg(feature = "handlers")]
 mod graphql_registry;
 #[cfg(feature = "handlers")]